fix: update uishell stories
